+++ node/src/tests/execution_requests.rs
@@ -1,0 +1,185 @@
+use crate::engine::Engine;
+use crate::test_harness::common;
+use crate::test_harness::common::get_default_engine_config;
+use crate::test_harness::mock_engine_client::MockEngineNetworkBuilder;
+use alloy_primitives::hex;
+use alloy_signer::k256::elliptic_curve::rand_core::OsRng;
+use commonware_cryptography::bls12381::dkg::ops;
+use commonware_cryptography::bls12381::primitives::variant::MinPk;
+use commonware_cryptography::{PrivateKeyExt, Signer};
+use commonware_macros::test_traced;
+use commonware_p2p::simulated;
+use commonware_p2p::simulated::{Link, Network};
+use commonware_runtime::deterministic::Runner;
+use commonware_runtime::{Clock, Metrics, Runner as _, deterministic};
+use commonware_utils::{from_hex_formatted, quorum};
+use std::collections::{HashMap, HashSet};
+use std::time::Duration;
+use summit_types::PrivateKey;
+use summit_types::execution_request::ExecutionRequest;
+
+#[test_traced]
+fn test_deposit_request() {
+    let n = 10;
+    let link = Link {
+        latency: 80.0,
+        jitter: 10.0,
+        success_rate: 0.98,
+    };
+    // Create context
+    let threshold = quorum(n);
+    let cfg = deterministic::Config::default().with_seed(0);
+    let executor = Runner::from(cfg);
+    executor.start(|context| async move {
+        // Create simulated network
+        let (network, mut oracle) = Network::new(
+            context.with_label("network"),
+            simulated::Config {
+                max_size: 1024 * 1024,
+            },
+        );
+
+        // Start network
+        network.start();
+
+        // Register participants
+        let mut signers = Vec::new();
+        let mut validators = Vec::new();
+        for i in 0..n {
+            let signer = PrivateKey::from_seed(i as u64);
+            let pk = signer.public_key();
+            signers.push(signer);
+            validators.push(pk);
+        }
+        validators.sort();
+        signers.sort_by_key(|s| s.public_key());
+        let mut registrations = common::register_validators(&mut oracle, &validators).await;
+
+        // Link all validators
+        common::link_validators(&mut oracle, &validators, link, None).await;
+
+        // Create the engine clients
+        let genesis_hash =
+            from_hex_formatted(common::GENESIS_HASH).expect("failed to decode genesis hash");
+        let genesis_hash: [u8; 32] = genesis_hash
+            .try_into()
+            .expect("failed to convert genesis hash");
+
+        // Create a single deposit request using the helper
+        let deposit_requests = common::create_deposit_requests(1);
+        let test_deposit = deposit_requests[0].clone();
+
+        // Convert to ExecutionRequest and then to Requests
+        let execution_requests = vec![ExecutionRequest::Deposit(test_deposit.clone())];
+        let requests = common::execution_requests_to_requests(execution_requests);
+
+        // Create execution requests map (add deposit to block 5)
+        let deposit_block_height = 5;
+        let stop_height = deposit_block_height + 1;
+        let mut execution_requests_map = HashMap::new();
+        execution_requests_map.insert(deposit_block_height, requests);
+
+        let engine_client_network = MockEngineNetworkBuilder::new(genesis_hash)
+            .with_execution_requests(execution_requests_map)
+            .build();
+
+        // Derive threshold
+        let (polynomial, shares) = ops::generate_shares::<_, MinPk>(&mut OsRng, None, n, threshold);
+
+        // Create instances
+        let mut public_keys = HashSet::new();
+        for (idx, signer) in signers.into_iter().enumerate() {
+            // Create signer context
+            let public_key = signer.public_key();
+            public_keys.insert(public_key.clone());
+
+            // Configure engine
+            let uid = format!("validator-{public_key}");
+            let namespace = String::from("_SEISMIC_BFT");
+
+            let engine_client = engine_client_network.create_client(uid.clone());
+
+            let config = get_default_engine_config(
+                engine_client,
+                uid.clone(),
+                genesis_hash,
+                namespace,
+                signer,
+                polynomial.clone(),
+                shares[idx].clone(),
+                validators.clone(),
+            );
+            let engine = Engine::new(
+                context.with_label(&uid),
+                config,
+                oracle.control(public_key.clone()),
+            )
+            .await;
+
+            // Get networking
+            let (pending, recovered, resolver, broadcast, backfill) =
+                registrations.remove(&public_key).unwrap();
+
+            // Start engine
+            engine.start(pending, recovered, resolver, broadcast, backfill);
+        }
+
+        // Poll metrics
+        let mut num_nodes_finished = 0;
+        loop {
+            let metrics = context.encode();
+
+            // Iterate over all lines
+            let mut success = false;
+            for line in metrics.lines() {
+                // Ensure it is a metrics line
+                if !line.starts_with("validator-") {
+                    continue;
+                }
+
+                // Split metric and value
+                let mut parts = line.split_whitespace();
+                let metric = parts.next().unwrap();
+                let value = parts.next().unwrap();
+
+                // If ends with peers_blocked, ensure it is zero
+                if metric.ends_with("_peers_blocked") {
+                    let value = value.parse::<u64>().unwrap();
+                    assert_eq!(value, 0);
+                }
+
+                if metric.ends_with("validator_balance") {
+                    let value = value.parse::<u64>().unwrap();
+                    // Parse the pubkey from the metric name using helper function
+                    if let Some(pubkey_hex) = common::parse_metric_substring(metric, "pubkey") {
+                        let bls_pubkey_hex = hex::encode(test_deposit.bls_pubkey);
+                        assert_eq!(bls_pubkey_hex, pubkey_hex);
+                        assert_eq!(value, test_deposit.amount);
+                        num_nodes_finished += 1;
+                        if num_nodes_finished == n {
+                            success = true;
+                            break;
+                        }
+                    } else {
+                        println!("{}: {} (failed to parse pubkey)", metric, value);
+                    }
+                }
+            }
+            if success {
+                break;
+            }
+
+            // Still waiting for all validators to complete
+            context.sleep(Duration::from_secs(1)).await;
+        }
+
+        // Check that all nodes have the same canonical chain
+        assert!(
+            engine_client_network
+                .verify_consensus(Some(stop_height))
+                .is_ok()
+        );
+
+        context.auditor().state()
+    })
+}
